+++ src/apis/workspace/usePostCreateWorkspaceUrl.ts
@@ -1,25 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
 import { axiosInstance } from '../axios.ts';
 import type {
-  CreateWorkspaceRequest,
-  CreateWorkspaceResponse,
   CreateWorkspaceUrlRequest,
   CreateWorkspaceUrlResponse,
 } from '../../types/workspace.ts';
-
-/**
- * 워크스페이스 생성 요청 함수
- */
-export const postCreateWorkspace = async (
-  data: CreateWorkspaceRequest
-): Promise<CreateWorkspaceResponse> => {
-  try {
-    const response = await axiosInstance.post<CreateWorkspaceResponse>('/api/workspace', data);
-    return response.data;
-  } catch (error) {
-    console.error('워크스페이스 생성 요청 실패:', error);
-    throw error;
-  }
-};
 
 /**
  * 워크스페이스 URL 생성 요청 함수
@@ -37,4 +21,10 @@ export const postCreateWorkspaceUrl = async (
     console.error('워크스페이스 URL 생성 요청 실패:', error);
     throw error;
   }
+};
+
+export const usePostCreateWorkspaceUrl = () => {
+  return useMutation<CreateWorkspaceUrlResponse, Error, CreateWorkspaceUrlRequest>({
+    mutationFn: postCreateWorkspaceUrl,
+  });
 };
